fix(game): F sweeps — all three classes closed; bg_panimate ushort promotion, bg_saga class_shader arm
F3: corrected sharp criterion (inclusive operators diverge in the
OPPOSITE rounding direction from strict — >= iff round-DOWN, <= iff
round-UP); mechanical scan of all 89 game files confirms every live
divergence-capable compare was already fixed in waves 2/3/5 and the
inclusive-divergent class is empty in the corpus; dead DebugLine
promoted for purity. F1 over the unaudited tail: q_math fully clean
(the float substrate), bg_lib linkage grounded (Raven's own unguarded
atof links natively, port matches); fixed bg_panimate numFrames-1 (C
promotes the ushort to signed int, 0 -> -1 into the dur<=1 fallback;
u16 port panicked/wrapped — product also runs in f64) and the bg_saga
class_shader missing-key arm (restore the error print, drop the forced
SPC_INFANTRY, stale-slot semantics now match). F2: zero guards lost
port-wide (956101f7 restores stand; ~90 params verified preserved);
two stale g_vehicles PORT-NOTEs corrected. ColorBytes3 §19 note,
setTable cite. Adversarially validated; full referee suite green.

+++ crates/mp/game/src/bg_panimate.rs
@@ -2458,9 +2458,11 @@ impl PmoveContext<'_> {
                         if setAnimFlags & SETANIM_FLAG_HOLDLESS != 0 {
                             // Make sure to only wait in full 1/20 sec server frame intervals.
                             // Oracle: `int dur; int speedDif;` — both truncate to int
-                            // before the add (bg_panimate.c:2839-2844).
-                            let dur0 = ((frame.numFrames - 1) as f32
-                                * (frame.frameLerp as f32).abs())
+                            // before the add; `numFrames-1` promotes the ushort to
+                            // SIGNED int (0 → -1, not a u16 wrap), and the product with
+                            // double `fabs` runs in f64 (bg_panimate.c:2839-2844).
+                            let dur0 = ((frame.numFrames as c_int - 1) as f64
+                                * ((frame.frameLerp as f32) as f64).abs())
                                 as c_int;
                             let speedDif = (dur0 as f32 - (dur0 as f32 * editAnimSpeed)) as c_int;
                             let dur = dur0 + speedDif;
@@ -2503,9 +2505,11 @@ impl PmoveContext<'_> {
                         if setAnimFlags & SETANIM_FLAG_HOLDLESS != 0 {
                             // Make sure to only wait in full 1/20 sec server frame intervals.
                             // Oracle: `int dur; int speedDif;` — both truncate to int
-                            // before the add (bg_panimate.c:2887-2892).
-                            let dur0 = ((frame.numFrames - 1) as f32
-                                * (frame.frameLerp as f32).abs())
+                            // before the add; `numFrames-1` promotes the ushort to
+                            // SIGNED int (0 → -1, not a u16 wrap), and the product with
+                            // double `fabs` runs in f64 (bg_panimate.c:2887-2892).
+                            let dur0 = ((frame.numFrames as c_int - 1) as f64
+                                * ((frame.frameLerp as f32) as f64).abs())
                                 as c_int;
                             let speedDif = (dur0 as f32 - (dur0 as f32 * editAnimSpeed)) as c_int;
                             let dur = dur0 + speedDif;
